feat(ui): extract EntitySortUtils, EntityPermissionUtils, EntityVoteUtils, EntityTagUtils, EntitySearchUtils, EntityColumnUtils, EntityBreadcrumbPureUtils, EntitySummaryPanelPureUtils from EntityPureUtils/EntityUtils/EntityBreadcrumbUtils/EntitySummaryPanelUtils
- Extract columnSorter/getColumnSorter → EntitySortUtils.ts
- Extract hasSchemaTab/hasLineageTab/hasCustomPropertiesTab/hasEditAccess → EntityPermissionUtils.ts
- Extract getEntityVoteStatus → EntityVoteUtils.ts
- Extract getEntityTags → EntityTagUtils.ts
- Extract searchInColumns/highlightEntityNameAndDescription/highlightSearchText/highlightSearchArrayElement → EntitySearchUtils.tsx
- Extract getEntityReferenceFromEntity/getEntityReferenceListFromEntities → EntityReferenceUtils.ts
- Extract getFrequentlyJoinedColumns → EntityColumnUtils.tsx (new file)
- Move all pure breadcrumb functions → EntityBreadcrumbPureUtils.ts; EntityBreadcrumbUtils.tsx becomes re-export facade
- Extract getSummaryListItemType/getSortedTagsWithHighlight/getMapOfListHighlights/getHighlightOfListItem/toEntityData → EntitySummaryPanelPureUtils.ts
- Extract filterItemsBySearchText/filterNestedFields → EntitySummaryPanelPureUtilsV1.ts (imported instead of redefined)
- Add backward-compat re-exports in all source files so existing imports continue to work
