+++ openmetadata-ui/src/main/resources/ui/src/utils/EntityColumnUtils.tsx
@@ -1,0 +1,102 @@
+/*
+ *  Copyright 2025 Collate.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+import { Popover } from 'antd';
+import { Fragment } from 'react';
+import { Link } from 'react-router-dom';
+import { FQN_SEPARATOR_CHAR } from '../constants/char.constants';
+import { EntityType, FqnPart } from '../enums/entity.enum';
+import type { ColumnJoins } from '../generated/entity/data/table';
+import {
+    checkIfJoinsAvailable,
+    getFrequentlyJoinedWithColumns
+} from './EntityPureUtils';
+import {
+    getPartialNameFromTableFQN,
+    getTableFQNFromColumnFQN
+} from './FqnUtils';
+import { getEntityDetailsPath } from './RouterUtils';
+
+export const getFrequentlyJoinedColumns = (
+  columnName: string,
+  joins: Array<ColumnJoins>,
+  columnLabel: string
+) => {
+  const frequentlyJoinedWithColumns = getFrequentlyJoinedWithColumns(
+    columnName,
+    joins
+  );
+
+  return checkIfJoinsAvailable(columnName, joins) ? (
+    <div className="m-t-sm" data-testid="frequently-joined-columns">
+      <span className="text-grey-muted m-r-xss">{columnLabel}:</span>
+      <span>
+        {frequentlyJoinedWithColumns.slice(0, 3).map((columnJoin, index) => (
+          <Fragment key={columnJoin.fullyQualifiedName}>
+            {index > 0 && <span className="m-r-xss">,</span>}
+            <Link
+              className="link-text"
+              to={getEntityDetailsPath(
+                EntityType.TABLE,
+                columnJoin.fullyQualifiedName
+              )}>
+              {getPartialNameFromTableFQN(
+                columnJoin.fullyQualifiedName,
+                [
+                  FqnPart.Database,
+                  FqnPart.Table,
+                  FqnPart.Schema,
+                  FqnPart.Column,
+                ],
+                FQN_SEPARATOR_CHAR
+              )}
+            </Link>
+          </Fragment>
+        ))}
+
+        {frequentlyJoinedWithColumns.length > 3 && (
+          <Popover
+            content={
+              <div className="text-left">
+                {frequentlyJoinedWithColumns?.slice(3).map((columnJoin) => (
+                  <Fragment key={columnJoin.fullyQualifiedName}>
+                    <a
+                      className="link-text d-block p-y-xss"
+                      href={getEntityDetailsPath(
+                        EntityType.TABLE,
+                        getTableFQNFromColumnFQN(
+                          columnJoin?.fullyQualifiedName
+                        ),
+                        getPartialNameFromTableFQN(
+                          columnJoin?.fullyQualifiedName,
+                          [FqnPart.Column]
+                        )
+                      )}>
+                      {getPartialNameFromTableFQN(
+                        columnJoin?.fullyQualifiedName,
+                        [FqnPart.Database, FqnPart.Table, FqnPart.Column]
+                      )}
+                    </a>
+                  </Fragment>
+                ))}
+              </div>
+            }
+            placement="bottom"
+            trigger="click">
+            <span className="show-more m-l-xss text-underline">...</span>
+          </Popover>
+        )}
+      </span>
+    </div>
+  ) : null;
+};
