+++ openmetadata-ui/src/main/resources/ui/src/utils/EntitySummaryPanelUtils.tsx
@@ -14,12 +14,11 @@
 
 import Icon from '@ant-design/icons';
 import { Col, Row, Typography } from 'antd';
-import { get, isEmpty, isUndefined } from 'lodash';
+import { get, isEmpty } from 'lodash';
 import { Link } from 'react-router-dom';
 import { SearchedDataProps } from '../../src/components/SearchedData/SearchedData.interface';
 import { ReactComponent as IconExternalLink } from '../assets/svg/external-links.svg';
 import { GenericProvider } from '../components/Customization/GenericProvider/GenericProvider';
-import { ColumnOrTask } from '../components/Database/ColumnDetailPanel/ColumnDetailPanel.interface';
 import SchemaEditor from '../components/Database/SchemaEditor/SchemaEditor';
 import APIEndpointSummary from '../components/Explore/EntitySummaryPanel/APIEndpointSummary/APIEndpointSummary';
 import { ColumnSummaryList } from '../components/Explore/EntitySummaryPanel/ColumnSummaryList/ColumnsSummaryList';
@@ -27,65 +26,75 @@ import DataProductSummary from '../components/Explore/EntitySummaryPanel/DataPro
 import DomainSummary from '../components/Explore/EntitySummaryPanel/DomainSummary/DomainSummary.component';
 import GlossaryTermSummary from '../components/Explore/EntitySummaryPanel/GlossaryTermSummary/GlossaryTermSummary.component';
 import SummaryList from '../components/Explore/EntitySummaryPanel/SummaryList/SummaryList.component';
-import {
-  BasicEntityInfo,
-  HighlightedTagLabel,
-} from '../components/Explore/EntitySummaryPanel/SummaryList/SummaryList.interface';
+import type { BasicEntityInfo } from '../components/Explore/EntitySummaryPanel/SummaryList/SummaryList.interface';
 import TagsSummary from '../components/Explore/EntitySummaryPanel/TagsSummary/TagsSummary.component';
 import MetricExpression from '../components/Metric/MetricExpression/MetricExpression';
 import RelatedMetrics from '../components/Metric/RelatedMetrics/RelatedMetrics';
 import { ICON_DIMENSION, NO_DATA_PLACEHOLDER } from '../constants/constants';
 import { CustomizeEntityType } from '../constants/Customize.constants';
-import { SummaryListHighlightKeys } from '../constants/EntitySummaryPanelUtils.constant';
 import { OperationPermission } from '../context/PermissionProvider/PermissionProvider.interface';
 import { CSMode } from '../enums/codemirror.enum';
 import { EntityType } from '../enums/entity.enum';
 import { SummaryEntityType } from '../enums/EntitySummary.enum';
-import { Tag } from '../generated/entity/classification/tag';
-import { APIEndpoint } from '../generated/entity/data/apiEndpoint';
-import { Chart } from '../generated/entity/data/chart';
-import { Container, TagLabel } from '../generated/entity/data/container';
-import { Dashboard } from '../generated/entity/data/dashboard';
-import { DashboardDataModel } from '../generated/entity/data/dashboardDataModel';
-import { Database } from '../generated/entity/data/database';
-import { GlossaryTerm } from '../generated/entity/data/glossaryTerm';
-import { Metric } from '../generated/entity/data/metric';
-import { MlFeature, Mlmodel } from '../generated/entity/data/mlmodel';
-import { Pipeline, Task } from '../generated/entity/data/pipeline';
-import { SearchIndex } from '../generated/entity/data/searchIndex';
-import {
+import type { Tag } from '../generated/entity/classification/tag';
+import type { APIEndpoint } from '../generated/entity/data/apiEndpoint';
+import type { Chart } from '../generated/entity/data/chart';
+import type { Container } from '../generated/entity/data/container';
+import type { Dashboard } from '../generated/entity/data/dashboard';
+import type { DashboardDataModel } from '../generated/entity/data/dashboardDataModel';
+import type { Database } from '../generated/entity/data/database';
+import type { GlossaryTerm } from '../generated/entity/data/glossaryTerm';
+import type { Metric } from '../generated/entity/data/metric';
+import type { MlFeature, Mlmodel } from '../generated/entity/data/mlmodel';
+import type { Pipeline, Task } from '../generated/entity/data/pipeline';
+import type { SearchIndex } from '../generated/entity/data/searchIndex';
+import type {
   StoredProcedure,
   StoredProcedureCodeObject,
 } from '../generated/entity/data/storedProcedure';
-import { Column, Table, TableConstraint } from '../generated/entity/data/table';
-import { Field, Topic } from '../generated/entity/data/topic';
-import { DataProduct } from '../generated/entity/domains/dataProduct';
-import { Domain } from '../generated/entity/domains/domain';
-import { EntityReference } from '../generated/tests/testCase';
-import { EntityData } from '../pages/TasksPage/TasksPage.interface';
+import type {
+  Column,
+  Table,
+  TableConstraint,
+} from '../generated/entity/data/table';
+import type { Field, Topic } from '../generated/entity/data/topic';
+import type { DataProduct } from '../generated/entity/domains/dataProduct';
+import type { Domain } from '../generated/entity/domains/domain';
+import type { EntityReference } from '../generated/tests/testCase';
+import {
+  getHighlightOfListItem,
+  getMapOfListHighlights,
+  getSummaryListItemType,
+  toEntityData,
+  type ListItemHighlights,
+  type SummaryListItem,
+} from './EntitySummaryPanelPureUtils';
 import entityUtilClassBase from './EntityUtilClassBase';
 import { getEntityName } from './EntityUtils';
 import { t } from './i18next/LocalUtil';
 import searchClassBase from './SearchClassBase';
 import { stringToHTML } from './StringUtils';
 
+export {
+  getHighlightOfListItem,
+  getMapOfListHighlights,
+  getSortedTagsWithHighlight,
+  getSummaryListItemType,
+  toEntityData,
+} from './EntitySummaryPanelPureUtils';
+export type { ListItemHighlights, SummaryListItem };
+export { toEntityData };
+
 const { Text } = Typography;
 
-export type SummaryListItem = Column | Field | Chart | Task | MlFeature;
-
-export interface ListItemHighlights {
-  highlightedTags?: BasicEntityInfo['tags'];
-  highlightedTitle?: string;
-  highlightedDescription?: string;
-}
-
-/* @param {
-    listItem: SummaryItem,
-    highlightedTitle: will be a string if the title of given summaryItem is present in highlights | undefined
-}
-
-    @return SummaryItemTitle
-*/
+/*
+ * @param {
+ *   listItem: SummaryItem,
+ *   highlightedTitle: will be a string if the title of given summaryItem is present in highlights | undefined
+ * }
+ *
+ *  @return SummaryItemTitle
+ */
 export const getTitle = (
   listItem: SummaryListItem,
   highlightedTitle?: ListItemHighlights['highlightedTitle']
@@ -134,174 +143,20 @@ export const getTitle = (
   );
 };
 
-/* @param {
-    entityType: will be any type of SummaryEntityType,
-    listItem: SummaryItem
-}
-    @return listItemType
-*/
-export const getSummaryListItemType = (
-  entityType: SummaryEntityType,
-  listItem: SummaryListItem
-): BasicEntityInfo['type'] => {
-  switch (entityType) {
-    case SummaryEntityType.COLUMN:
-    case SummaryEntityType.FIELD:
-    case SummaryEntityType.MLFEATURE:
-    case SummaryEntityType.SCHEMAFIELD:
-      return (listItem as Column | Field | MlFeature).dataType;
-    case SummaryEntityType.CHART:
-      return (listItem as Chart).chartType;
-    case SummaryEntityType.TASK:
-      return (listItem as Task).taskType;
-    default:
-      return '';
-  }
-};
-
 /*
-    @params {
-        sortTagsBasedOnGivenTagFQNs: array of TagFQNs,
-        tags: Tags array,
-    }
-
-    @return array of tags highlighted and sorted if tagFQN present in sortTagsBasedOnGivenTagFQNs
-*/
-export const getSortedTagsWithHighlight = (
-  tags: TagLabel[] = [],
-  sortTagsBasedOnGivenTagFQNs: string[] = []
-): ListItemHighlights['highlightedTags'] => {
-  const { sortedTags, remainingTags } = tags.reduce(
-    (acc, tag) => {
-      if (sortTagsBasedOnGivenTagFQNs.includes(tag.tagFQN)) {
-        acc.sortedTags.push({ ...tag, isHighlighted: true });
-      } else {
-        acc.remainingTags.push(tag);
-      }
-
-      return acc;
-    },
-    {
-      sortedTags: [] as HighlightedTagLabel[],
-      remainingTags: [] as TagLabel[],
-    }
-  );
-
-  return [...sortedTags, ...remainingTags];
-};
-
-/* 
-    @param {highlights: all the other highlights come from the query api
-        only omitted displayName and description key as it is already updated in parent component
-    }
-
-    @return {
-        listHighlights: single array of all highlights get from query api 
-        listHighlightsMap: to reduce the search time complexity in listHighlight
-    }
-
-    Todo: apply highlights on entityData in parent where we apply highlight for entityDisplayName and entityDescription
-    for that we need to update multiple summary components
-*/
-export const getMapOfListHighlights = (
-  highlights?: SearchedDataProps['data'][number]['highlight']
-): {
-  listHighlights: string[];
-  listHighlightsMap: { [key: string]: number };
-} => {
-  // checking for the all highlight key present in highlight get from query api
-  // and create a array of highlights
-  const listHighlights: string[] = [];
-  SummaryListHighlightKeys.forEach((highlightKey) => {
-    listHighlights.push(...get(highlights, highlightKey, []));
-  });
-
-  // using hashmap methodology to reduce the search time complexity from O(n) to O(1)
-  // to get highlight from the listHighlights array for applying highlight
-  const listHighlightsMap: { [key: string]: number } = {};
-
-  listHighlights?.reduce((acc, colHighlight, index) => {
-    acc[colHighlight.replaceAll(/<\/?span(.*?)>/g, '')] = index;
-
-    return acc;
-  }, listHighlightsMap);
-
-  return { listHighlights, listHighlightsMap };
-};
-
-/*
-    @params {
-        listItem: SummaryItem
-        tagsHighlights: tagFQNs array to highlight and sort tags
-        listHighlights: single array of all highlights get from query api 
-        listHighlightsMap: to reduce the search time complexity in listHighlight
-    }
-    @return highlights of listItem
-*/
-export const getHighlightOfListItem = (
-  listItem: SummaryListItem,
-  tagHighlights: string[],
-  listHighlights: string[],
-  listHighlightsMap: { [key: string]: number }
-): ListItemHighlights => {
-  let highlightedTags;
-  let highlightedTitle;
-  let highlightedDescription;
-
-  // if any of the listItem.tags present in given tagHighlights list then sort and highlights the tag
-  const shouldSortListItemTags = listItem.tags?.find((tag) =>
-    tagHighlights.includes(tag.tagFQN)
-  );
-
-  if (shouldSortListItemTags) {
-    highlightedTags = getSortedTagsWithHighlight(listItem.tags, tagHighlights);
-  }
-
-  // highlightedListItemNameIndex will be undefined if listItem.name is not present in highlights
-  const highlightedListItemNameIndex = listHighlightsMap[listItem.name ?? ''];
-
-  const shouldApplyHighlightOnTitle = !isUndefined(
-    highlightedListItemNameIndex
-  );
-
-  if (shouldApplyHighlightOnTitle) {
-    highlightedTitle = listHighlights[highlightedListItemNameIndex];
-  }
-
-  // highlightedListItemDescriptionIndex will be undefined if listItem.description is not present in highlights
-  const highlightedListItemDescriptionIndex =
-    listHighlightsMap[listItem.description ?? ''];
-
-  const shouldApplyHighlightOnDescription = !isUndefined(
-    highlightedListItemDescriptionIndex
-  );
-
-  if (shouldApplyHighlightOnDescription) {
-    highlightedDescription =
-      listHighlights[highlightedListItemDescriptionIndex];
-  }
-
-  return {
-    highlightedTags,
-    highlightedTitle,
-    highlightedDescription,
-  };
-};
-
-/*
-    @params {
-        entityType: SummaryEntityType,
-        entityInfo: Array<SummaryListItem> = [],
-        highlights: highlights get from the query api + highlights added for tags (i.e. tag.name)
-        tableConstraints: only pass for SummayEntityType.Column
-    }
-    @return sorted and highlighted listItem array, but listItem will be type of BasicEntityInfo
-    
-    Note: SummaryItem will be sort and highlight only if -
-        # if listItem.tags present in highlights.tags
-        # if listItem.name present in highlights comes from query api
-        # if listItem.description present in highlights comes from query api
-*/
+ *  @params {
+ *     entityType: SummaryEntityType,
+ *     entityInfo: Array<SummaryListItem> = [],
+ *     highlights: highlights get from the query api + highlights added for tags (i.e. tag.name)
+ *     tableConstraints: only pass for SummayEntityType.Column
+ *  }
+ *  @return sorted and highlighted listItem array, but listItem will be type of BasicEntityInfo
+ *
+ *  Note: SummaryItem will be sort and highlight only if -
+ *      # if listItem.tags present in highlights.tags
+ *      # if listItem.name present in highlights comes from query api
+ *      # if listItem.description present in highlights comes from query api
+ */
 export const getFormattedEntityData = (
   entityType: SummaryEntityType,
   entityInfo: Array<SummaryListItem> = [],
@@ -312,13 +167,9 @@ export const getFormattedEntityData = (
     return [];
   }
 
-  // Only go ahead if entityType is present in SummaryEntityType enum
   if (Object.values(SummaryEntityType).includes(entityType)) {
-    // tagHighlights is the array of tagFQNs for highlighting tags
     const tagHighlights = get(highlights, 'tag.name', [] as string[]);
 
-    // listHighlights i.e. highlight get from query api
-    // listHighlightsMap i.e. map of highlight get from api to reduce search time complexity in highlights array
     const { listHighlights, listHighlightsMap } =
       getMapOfListHighlights(highlights);
 
@@ -330,7 +181,6 @@ export const getFormattedEntityData = (
 
     const { highlightedListItem, remainingListItem } = entityInfo.reduce(
       (acc, listItem) => {
-        // return the highlight of listItem
         const { highlightedTags, highlightedTitle, highlightedDescription } =
           getHighlightOfListItem(
             listItem,
@@ -339,7 +189,6 @@ export const getFormattedEntityData = (
             listHighlightsMap
           );
 
-        // convert listItem in BasicEntityInfo type
         const listItemModifiedData = {
           name: listItem.name ?? '',
           title: getTitle(listItem, highlightedTitle),
@@ -362,7 +211,6 @@ export const getFormattedEntityData = (
           }),
         };
 
-        // if highlights present in listItem then sort the listItem
         if (highlightedTags || highlightedTitle || highlightedDescription) {
           acc.highlightedListItem.push(listItemModifiedData);
         } else {
@@ -681,35 +529,4 @@ export const getEntityChildDetails = (
       <Col span={24}>{childComponent}</Col>
     </Row>
   );
-};
-
-/**
- * Convert ColumnOrTask to EntityData for CustomPropertiesSection
- * @param column - Column or task-like entity
- * @returns EntityData object with extension property if available, or undefined
- */
-export const toEntityData = (
-  column: ColumnOrTask | null
-): EntityData | undefined => {
-  if (!column) {
-    return undefined;
-  }
-
-  // Check if column has extension property and create a new object that satisfies EntityData
-  // extension is typed as 'any' in entity interfaces, but EntityData.extension expects Record<string, unknown>
-  // so we cast it after runtime validation to ensure type safety
-  const extension =
-    'extension' in column &&
-    typeof column.extension === 'object' &&
-    column.extension !== null
-      ? column.extension
-      : undefined;
-
-  // Create an object that satisfies EntityData interface with index signature
-  const entityData: EntityData = {} as EntityData;
-  if (extension) {
-    entityData.extension = extension;
-  }
-
-  return entityData;
 };
