+++ openmetadata-ui/src/main/resources/ui/src/utils/EntityVoteUtils.ts
@@ -1,0 +1,39 @@
+/*
+ *  Copyright 2025 Collate.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+import { isUndefined } from 'lodash';
+import { QueryVoteType } from '../components/Database/TableQueries/TableQueries.interface';
+import { Votes } from '../generated/type/votes';
+
+/**
+ * Take entity vote and userId as input and return name for vote status type
+ * @param votes - entity votes
+ * @param userId - current user id
+ * @returns - vote status type
+ */
+export const getEntityVoteStatus = (userId: string, votes?: Votes) => {
+  if (isUndefined(votes)) {
+    return QueryVoteType.unVoted;
+  }
+
+  const upVoters = votes.upVoters ?? [];
+  const downVoters = votes.downVoters ?? [];
+
+  if (upVoters.some((user) => user.id === userId)) {
+    return QueryVoteType.votedUp;
+  } else if (downVoters.some((user) => user.id === userId)) {
+    return QueryVoteType.votedDown;
+  } else {
+    return QueryVoteType.unVoted;
+  }
+};
